Fix relativo localizzazione utente mappa
Parametri in float invece che in integer

+++ app/controllers/monuments_controller.rb
@@ -6,7 +6,7 @@ class MonumentsController < ApplicationController
     if params[:latitude] && params[:longitude]
       @pagy, @monument = pagy(Monument.near([BigDecimal.new(params[:latitude]), BigDecimal.new(params[:longitude])]))
       @monument_nopagy = Monument.near([BigDecimal.new(params[:latitude]), BigDecimal.new(params[:longitude])])
-      @geocenter = [params[:latitude].to_i, params[:longitude].to_i]
+      @geocenter = [params[:latitude].to_f, params[:longitude].to_f]
     elsif params[:city]
       @pagy, @monument = pagy(Monument.near("#{params[:city]}"))
       @monument_nopagy = Monument.near(params[:city])
